+++ acfml.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name: ACF Multilingual
- * Version: 1.1.6
+ * Version: 1.1.7
  * Author: Rasso Hilber
  * Description: A lightweight solution to support multiple languages with WordPress and Advanced Custom Fields
  * Author URI: https://rassohilber.com
@@ -19,7 +19,6 @@ define( 'ACFML_PATH', plugin_dir_path( __FILE__ ) );
 define( 'ACFML_BASENAME', plugin_basename( __FILE__ ) );
 define( 'ACFML_URL', plugins_url('/', __FILE__) );
 
-require_once(ACFML_PATH . 'vendor/autoload.php');
 require_once(ACFML_PATH . 'api.php');
 
 /**