+++ changelog.md
@@ -1,6 +1,7 @@
 #### 1.1.6 (2021-11-29)
 
-- Make running test:php easier (#0ce3ec1)
+- Prepare for testing new config (#d3b308c)
+- Make running test:php easier (#d16f761)
 - Allow booleans for post types and taxonomies in `acfml.config.json` (#2a61e12)
 
 #### 1.1.5 (2021-11-19)